feat: 특정 데이터 구역 클릭 시 동일한 ID 값을 가진 데이터 구역 하이라이트
