--- src/interface/props.d.ts
@@ -1,6 +1,0 @@
-import { IChart } from './chartData';
-
-export interface IChartProps {
-  data: IChart[];
-  filteredData: string[];
-}
